requirements.txt
colorama==0.3.7
py==1.4.31
pytest==3.0.5
PyYAML==3.12
requests==2.12.3
